refactor: code split octokit and worker code
